+++ CHANGELOG.md
@@ -6,6 +6,22 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.1.0-alpha.12] · 2023-??-??
+[0.1.0-alpha.12]: /../../tree/v0.1.0-alpha.12
+
+[Diff](/../../compare/v0.1.0-alpha.11...v0.1.0-alpha.12) | [Milestone](/../../milestone/13)
+
+### Changed
+
+- UI:
+    - Download and Save as option from context menu ([#697], [#654])
+
+[#654]: /../../issues/654
+[#697]: /../../pull/697
+
+
+
+
 ## [0.1.0-alpha.11] · 2023-11-02
 [0.1.0-alpha.11]: /../../tree/v0.1.0-alpha.11
 
@@ -21,7 +37,6 @@ All user visible changes to this project will be documented in this file. This p
         - Display message field while loading. ([#662], [#634])
         - Display small images smaller. ([#688], [#653])
         - Message attachments and replies editing. ([#671], [#557])
-        - Download and Save as option from context menu ([#697], [#654]) 
     - Disabled larger fonts accessibility setting temporary. ([#679])
     - Home page:
         - Redesigned introduction modal. ([#668], [#633])
@@ -41,7 +56,6 @@ All user visible changes to this project will be documented in this file. This p
 [#634]: /../../issues/634
 [#633]: /../../issues/633
 [#653]: /../../issues/653
-[#654]: /../../issues/654
 [#662]: /../../pull/662
 [#663]: /../../pull/663
 [#668]: /../../pull/668
@@ -49,7 +63,6 @@ All user visible changes to this project will be documented in this file. This p
 [#672]: /../../pull/672
 [#679]: /../../pull/679
 [#688]: /../../pull/688
-[#697]: /../../pull/697